requirements.txt
celery==5.2.7
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
psycopg2==2.9.5
redis==4.5.1
requests==2.28.2
retry==0.9.2
result==0.9.0
django-extensions==3.2.1